requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.1.1
cycler==0.12.1
databricks-sdk==0.36.0
databricks-vectorsearch==0.40
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.2.17
langchain-core==0.2.43
langchain-databricks==0.1.0rc1
langchain-text-splitters==0.2.4
langsmith==0.1.142
Mako==1.3.6
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mlflow==2.17.2
mlflow-skinny==2.17.2
multidict==6.1.0
numpy==1.24.4
openai==1.54.3
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
orjson==3.10.11
packaging==24.2
pandas==2.0.3
pillow==10.4.0
propcache==0.2.0
protobuf==4.25.5
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.9.2
pydantic-core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
tenacity==8.5.0
threadpoolctl==3.5.0
tqdm==4.67.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
werkzeug==3.0.6
wrapt==1.16.0
yarl==1.15.2
zipp==3.20.2
